CHORE #264 : SNS 연동 관련 provider 패키지로 이동
