+++ src/main/java/com/lokoko/global/auth/controller/enums/ResponseMessage.java
@@ -11,8 +11,6 @@ public enum ResponseMessage {
     URL_GET_SUCCESS("리다이렉트 URL 조회에 성공했습니다."),
     REFRESH_TOKEN_REISSUE("리프레시 토큰 재발급에 성공했습니다."),
     ROLE_ASSIGNED_SUCCESS("역할 설정이 완료되었습니다."),
-    TIKTOK_CONNECT_SUCCESS("틱톡 계정 연결에 성공했습니다."),
-    TIKTOK_REDIRECT_URI_GET_SUCCESS("틱톡 redirect uri 반환에 성공했습니다"),
     GET_LOGIN_USER_ID_SUCCESS("로그인한 유저의 ID를 성공적으로 불러왔습니다.");
 
     private final String message;
