--- src/main/java/com/lokoko/global/auth/google/dto/GoogleTokenDto.java
@@ -1,15 +1,0 @@
-package com.lokoko.global.auth.google.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record GoogleTokenDto(
-        String accessToken,
-        String refreshToken,
-        Long expiresIn,
-        String idToken,
-        String tokenType,
-        String scope
-) {
-}
