--- src/main/java/com/lokoko/global/auth/line/dto/LineProfileDto.java
@@ -1,10 +1,0 @@
-package com.lokoko.global.auth.line.dto;
-
-public record LineProfileDto(
-        String userId,
-        String displayName,
-        String pictureUrl,
-        String statusMessage,
-        String email
-) {
-}
