--- src/main/java/com/lokoko/global/auth/line/dto/LineUserInfoDto.java
@@ -1,8 +1,0 @@
-package com.lokoko.global.auth.line.dto;
-
-public record LineUserInfoDto(
-        String sub,
-        String name,
-        String picture
-) {
-}
