+++ src/main/java/com/lokoko/global/auth/provider/insta/usecase/InstaConnectionUsecase.java
@@ -1,0 +1,11 @@
+package com.lokoko.global.auth.provider.insta.usecase;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class InstaConnectionUsecase {
+}
