+++ src/main/java/com/lokoko/global/auth/provider/line/dto/LineProfileDto.java
@@ -1,0 +1,10 @@
+package com.lokoko.global.auth.provider.line.dto;
+
+public record LineProfileDto(
+        String userId,
+        String displayName,
+        String pictureUrl,
+        String statusMessage,
+        String email
+) {
+}
