+++ src/main/java/com/lokoko/global/auth/provider/line/dto/LineUserInfoDto.java
@@ -1,0 +1,8 @@
+package com.lokoko.global.auth.provider.line.dto;
+
+public record LineUserInfoDto(
+        String sub,
+        String name,
+        String picture
+) {
+}
