+++ src/main/java/com/lokoko/global/auth/provider/tiktok/TikTokProperties.java
@@ -1,0 +1,13 @@
+package com.lokoko.global.auth.provider.tiktok;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "tiktok")
+public record TikTokProperties(
+        String clientKey,
+        String clientSecret,
+        String redirectUri,
+        String baseUrl,
+        String scope
+) {
+}
