+++ src/main/java/com/lokoko/global/auth/provider/tiktok/dto/TikTokErrorDto.java
@@ -1,0 +1,12 @@
+package com.lokoko.global.auth.provider.tiktok.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record TikTokErrorDto(
+        String code,
+        String message,
+        String logId
+) {
+}
