+++ src/main/java/com/lokoko/global/auth/provider/tiktok/dto/TikTokTokenDto.java
@@ -1,0 +1,16 @@
+package com.lokoko.global.auth.provider.tiktok.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record TikTokTokenDto(
+        String accessToken,
+        String refreshToken,
+        Long expiresIn,
+        Long refreshExpiresIn,
+        String tokenType,
+        String scope,
+        String openId
+) {
+}
