--- src/main/java/com/lokoko/global/auth/tiktok/TikTokProperties.java
@@ -1,13 +1,0 @@
-package com.lokoko.global.auth.tiktok;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "tiktok")
-public record TikTokProperties(
-        String clientKey,
-        String clientSecret,
-        String redirectUri,
-        String baseUrl,
-        String scope
-) {
-}
