--- src/main/java/com/lokoko/global/auth/tiktok/dto/TikTokTokenDto.java
@@ -1,16 +1,0 @@
-package com.lokoko.global.auth.tiktok.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record TikTokTokenDto(
-        String accessToken,
-        String refreshToken,
-        Long expiresIn,
-        Long refreshExpiresIn,
-        String tokenType,
-        String scope,
-        String openId
-) {
-}
