--- src/main/java/com/lokoko/global/auth/tiktok/dto/TikTokUserInfoResponse.java
@@ -1,16 +1,0 @@
-package com.lokoko.global.auth.tiktok.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record TikTokUserInfoResponse(
-        TikTokDataWrapper data,
-        TikTokErrorDto error
-) {
-    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-    public record TikTokDataWrapper(
-            TikTokProfileDto user
-    ) {
-    }
-}
